refactor:cloudDataHandlerをリファクタリング
