+++ src/constants/config.ts
@@ -69,6 +69,8 @@ export const CONFIG = {
   AWS: {
     /** デフォルトリージョン */
     DEFAULT_REGION: "auto",
+    /** 削除バッチのサイズ（S3の制限による）*/
+    DELETE_BATCH_SIZE: 1000,
     /** リクエストタイムアウト (ms) */
     REQUEST_TIMEOUT_MS: 30000,
     /** オブジェクト一覧取得の最大反復回数（無限ループ防止） */
